requirements.txt
alembic==0.8.7
bcrypt==3.1.0
blinker==1.4
cffi==1.7.0
click==6.6
dnspython==1.14.0
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
pycparser==2.14
PyMySQL==0.7.6
python-editor==1.0.1
six==1.10.0
SQLAlchemy==1.0.14
Werkzeug==0.11.10
WTForms==2.1
